fix: takeout separated index service of product module
