+++ providers/AppServiceInjectionProvider.cs
@@ -7,7 +7,6 @@ namespace onboarding_dotnet.Providers
         public void InitServiceInjection()
         {
             builder.Services.AddScoped<CategoryService, CategoryService>();
-            builder.Services.AddScoped<ProductIndexService, ProductIndexService>();
             builder.Services.AddScoped<ProductService, ProductService>();
             builder.Services.AddScoped<OrderService, OrderService>();
             builder.Services.AddScoped<TransactionService, TransactionService>();
